Enhance: Improve Game Flow Handling
The main change here is refactoring game flow. Previously the server set a timeout for roughly
the duration of the BattlePhase until it then ends a BattlePhase. This was very unreliable for obvious reasons.
The now improved implementation has the client notify the server that the animation is finished, to which the server then
forwards him to the next phase.